Foremka added

+++ Darwin/src/main/scala/pl/poznan/put/darwin/model/Config.scala
@@ -35,7 +35,7 @@ class Config(parser: ConfigParser) {
   /*
    * DOMLEM CONFIG
    */
-  val DOMLEM_CONFIDECE_LEVEL = parser.getInt("domlem", "confidencelevel")
+  val DOMLEM_CONFIDECE_LEVEL = parser.getDouble("domlem", "confidencelevel")
   val CONDITION_SELECTION_METHOD = VCDomLem.MIX_CONDITIONS_FROM_DIFFERENT_OBJECTS
   val NEGATIVE_EXAMPLES_TREATMENT = VCDomLem.COVER_ONLY_INCONSISTENT_AND_BOUNDARY_NEGATIVE_EXAMPLES
